Fix broken ir macro injection
The "inject" procedure is not intended for removing syntax from the
input pattern; it is intended to prevent the *output* from being
renamed after the fact.

If the input symbol is to be taken as-is, use strip-syntax which will
do exactly that.

So: input => strip-syntax => combine symbol names => inject => output

Signed-off-by: arthurmaciel <arthurmaciel@gmail.com>